+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Optimize `CTRFileIO` to re-use existing cipher object when possible (seeking invalidates the current one)
 * Optimize `RomFSReader` by reading entire directory and file metadata at once before traversing, significantly reducing the amount of read calls to the underlying file
 * Optimize `RomFSReader` to reduce the read calls for the header (once for raw lv3, twice for IVFC)
+* Check for unformatted saves in `DISA` (the first 0x20 bytes are all NULL and the rest is garbage)
 
 ## v0.6.0 - January 26, 2022
 ### Highlights